ISSUE #3705 - moved image cahnge button to a seperate file
